requirements.txt
Flask==2.3.3
ryanair==2.5.1
requests==2.31.0
python-dotenv==1.0.0
amadeus==8.0.0